Added the x64 platform to the development solution. Changed the dll-keyword-handling. Added PASTEL_DYNAMIC_LIBRARIES and PASTEL_LARGE_INTEGER. Changed the definition of 'integer'. Updated the build system. Updated SecondIterator and SparseIterator. Updated Rational class to work with different kinds of native integers.

+++ pastel/gfx/draw_perspective_triangle.hpp
@@ -462,7 +462,7 @@ namespace Pastel
 
 		for (integer y = yMin;y < yMid;++y)
 		{
-			const integer xBegin = std::max(toPixelSpanPoint(xLeft), 0);
+			const integer xBegin = std::max(toPixelSpanPoint(xLeft), (integer)0);
 			const integer xEnd = std::min(toPixelSpanPoint(xRight), width);
 
 			if (xEnd - xBegin > 0)
@@ -575,7 +575,7 @@ namespace Pastel
 
 		for (integer y = yMid;y < yMax;++y)
 		{
-			const integer xBegin = std::max(toPixelSpanPoint(xLeft), 0);
+			const integer xBegin = std::max(toPixelSpanPoint(xLeft), (integer)0);
 			const integer xEnd = std::min(toPixelSpanPoint(xRight), width);
 
 			if (xEnd - xBegin > 0)
